// src/app/models/admin/capphephoatdongkhoangsan/cptanthukhoangsan/cptanthutoadokhuvuc.model.ts
export class OutputCpTanThuToaDoKhuVucModel {
  idtanthutoadokhuvuc: string;
  sohieu: string;
  toadox: string;
  toadoy: string;
  idcappheptanthu: string;
  idtanthukhuvuc: string;
  loaikhuvuc: number;
  loaicapphep: number;
  siteid: string;
  thutu: number;
}

export class InputCpTanThuToaDoKhuVucModel {
  idtanthutoadokhuvuc: string;
  sohieu: string;
  toadox: string;
  toadoy: string;
  idcappheptanthu: string;
  idtanthukhuvuc: string;
  loaicapphep: number;
  loaikhuvuc: number;
  thutu: number;
}